pyproject.toml
[project]
name = "alpaca_trading_bot"
version = "0.1.0"
description = "Ein Trading-Bot für Alpaca mit Rolling Regression Strategie und Hebel"
authors = [
    { name = "Meric Kaynak", email = "mmk23.business@gmail.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"

dependencies = [
    "pandas>=2.0",
    "numpy>=1.25",
    "yfinance>=0.2.26",
    "alpaca-py==0.43.2",
    "toml>=0.10.2",
    "schedule>=1.2.0", # NEU für das tägliche Ausführen
    "pytz>=2023.3",     # NEU für CET Zeitzone
    "pydantic>=2.12.5",
    "python-dotenv>=1.0.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"